LayoutTests/http/tests/inspector/resource-tree/resource-tree-no-xhrs-expected.txt

CONSOLE MESSAGE: line 13: Done.
Tests that XHRs are not added to resourceTreeModel.

Bug 60321
